refactor(router): split bible.rs into router/bible/ submodules (#590)
router/bible.rs was at 885 prod lines, heading into the 1000-line hard
fail. Convert it to a router/bible/ directory, same pattern already used by
router/integrations/ (resolume, android_stage, ableset, osc, video_source):
one handler-group file per concern, pub(crate) items so router.rs's flat
route table (unchanged in structure) can reference them by qualified path.

- bible/translations.rs — translation admin + import/refresh trigger
- bible/browse.rs — book/chapter listing, search, single-passage lookup
- bible/broadcast.rs — active-broadcast readback + reference/slide triggers
  + clear + preferences
- bible/resolve.rs — BibleSlideDto + conversions + /bible/resolve preview
  (shared with presentations.rs)
- bible/presentations.rs — presentation CRUD + slide update/append/
  delete/reorder

router.rs's route table keeps the same paths and handler names, only
qualified by submodule. The one cross-reference outside bible.rs itself —
router/tests.rs's `use crate::router::bible::BibleImportSummaryDto` — is
updated to the new `bible::translations::` path.

Pure code movement, no behavior change, no route changes. All six new
files land well under the 800-line warn cap. Existing test suite is the
safety net (only the one import path adjusted).

Closes #590
